backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
google-genai>=1.61.0
httpx==0.28.0
python-multipart==0.0.20
aiofiles==24.1.0
